store/tikv:remove "tidb/tablecodec"  from snapshot.go (#23686)
