+++ server/conn_test.go
@@ -34,7 +34,6 @@ import (
 	"github.com/pingcap/tidb/sessionctx/variable"
 	"github.com/pingcap/tidb/store/mockstore"
 	"github.com/pingcap/tidb/store/mockstore/unistore"
-	"github.com/pingcap/tidb/store/tikv"
 	"github.com/pingcap/tidb/store/tikv/mockstore/cluster"
 	"github.com/pingcap/tidb/table"
 	"github.com/pingcap/tidb/util/arena"
@@ -689,6 +688,10 @@ func (ts *ConnTestSuite) TestShutdownOrNotify(c *C) {
 	c.Assert(cc.status, Equals, connStatusWaitShutdown)
 }
 
+type snapshotCache interface {
+	SnapCacheHitCount() int
+}
+
 func (ts *ConnTestSuite) TestPrefetchPointKeys(c *C) {
 	cc := &clientConn{
 		alloc: arena.NewAllocator(1024),
@@ -718,7 +721,7 @@ func (ts *ConnTestSuite) TestPrefetchPointKeys(c *C) {
 	c.Assert(err, IsNil)
 	c.Assert(txn.Valid(), IsTrue)
 	snap := txn.GetSnapshot()
-	c.Assert(snap.(*tikv.KVSnapshot).SnapCacheHitCount(), Equals, 4)
+	c.Assert(snap.(snapshotCache).SnapCacheHitCount(), Equals, 4)
 	tk.MustExec("commit")
 	tk.MustQuery("select * from prefetch").Check(testkit.Rows("1 1 2", "2 2 4", "3 3 4"))
